requirements.txt
kubernetes>=11.0.0
Flask>=1.1
flask-bootstrap4>=4.0
flask-nav>=0.3
Flask-HTMLmin>=1.5